feat (list): removing sleep time [ICTRAINING-1780]
