+++ lists/templates/home.html
@@ -12,7 +12,7 @@
     </form>
     <table id="id_list_table">
       {% for item in items %}
-        <tr><td>{{ forloop.counter }}: {{ new_item_text }}</td></tr>
+        <tr><td>{{ forloop.counter }}: {{ item.text }}</td></tr>
       {% endfor %}
     </table>
   </body>
